Make lookup in `.compile` file case-insensitive
I had an Xcode project where the case of filenames displayed in Xcode was not
consistent with the case of files on disk (e.g. `MyClass.swift` vs
`myclass.swift`). Techincally, these filenames are the same since MacOS's
file system is case-insensitive. However, because of these discrepancies,
`xcode-build-server` was not able to retrieve the correct compile command for
all files.